Both queue and and realtime taskqueues are working

+++ src/fi/swarco/dataHandling/MakeSendJsonOperations.java
@@ -14,6 +14,15 @@ public class MakeSendJsonOperations {
     public static void setjSonPermanentData(String pJSonPermanentData) {
         jSonPermanentData = pJSonPermanentData;
     }
+
+    private static int intSleep =5000;
+    public static int getSleep() {
+        return intSleep;
+    }
+    public static void setSleep(int pSleep) {
+         intSleep=pSleep;
+    }
+
     private static String jSonMeasurementData = "novalue";
     public static String getJSonMeasurementData() {
         return jSonMeasurementData;
@@ -72,6 +81,7 @@ public class MakeSendJsonOperations {
     public int PollOfWorks() {
         // check is ther work is if it is transfer them to work queue
         int iloop = 1;
+        int intSleep=0;
         int iRet = INT_RET_NOT_OK;
         iRet = MeasurementTaskHandling.MakeConnection(SwarcoEnumerations.ConnectionType.SQLSERVER_LOCAL_JOMNIATEST);
         if (iRet != INT_RET_OK) {
@@ -85,9 +95,6 @@ public class MakeSendJsonOperations {
                 return THERE_IS_WORK;
             }
             while (iloop == 1) {
-// 2. Controllerworks
-// 3. Detectorworks
-// 4. Measurementsworks
                 if (th.AnyWorkIntersection() > 0) {
                     iRet = th.TransferIntersectionTasksToWorkQueue();
                     if (iRet != TASK_TRANSFER_OK) {
@@ -168,35 +175,28 @@ public class MakeSendJsonOperations {
                     return THERE_IS_WORK;
                 }
                 if (th.AnyWorkMeasurements() == 0) {
-                    Thread.sleep(5000);   // 5 seconds sleep
+                    Thread.sleep(getSleep());   // 5 seconds sleep
                 } else {
                     // transfer tasks to work table
-                  iRet = th.TransferMeasurementTasksToWorkQueue();
-                  if (iRet != TASK_TRANSFER_OK) {
-                     logger.info("Task Transfer error iRet = " + iRet);
-                      return iRet;
-                  }
+                    iRet=  th.DeleteTrashTasksBeforeHand();
+                    if (iRet != DELETE_TRASH_TASK_OK) {
+                        logger.info("Task Transfer error iRet = " + iRet);
+                        return iRet;
+                    }
+                    iRet = th.TransferMeasurementTasksToWorkQueue();
+                     if (iRet != TASK_TRANSFER_OK) {
+                        logger.info("Task Transfer error iRet = " + iRet);
+                        return iRet;
+                    }
                 // update work from super
-                  iRet = th.FillUpTasks();
-                   if (iRet != FILL_UP_TASK_OK) {
-                      logger.info("Fill UpTask error iRet = " + iRet);
-                      return iRet;
-                   }
-                // because Intersection or controller is not visible
-                   iRet = th.DeleteNotFilledTasks();
-                   if (iRet != DELETE_UNFILLABLE_TASK_OK) {
-                      logger.info("Fill UpTask error iRet = " + iRet);
-                      return iRet;
-                   }
-                // delete tasks from task table
-                   iRet=th.DeleteTasksFromDb();
-                    if (iRet < 0) {
-                       logger.info("Task Delete error iRet = " + iRet);
+                    iRet = th.FillUpTasks();
+                    if (iRet != FILL_UP_TASK_OK) {
+                       logger.info("Fill UpTask error iRet = " + iRet);
                        return iRet;
                     }
-               }
-            return THERE_IS_WORK;
-        }
+                }
+               return THERE_IS_WORK;
+            }
         } catch (Exception e) {
             logger.info(ExceptionUtils.getRootCauseMessage(e));
             logger.info(ExceptionUtils.getFullStackTrace(e));
@@ -214,16 +214,25 @@ public class MakeSendJsonOperations {
         TRPXMeasurementTaskData ce = null;
         int iRound = 0;
         int iloop = 1;
+        int iRet2 =0;
         String strHelp1 = NO_VALUE;
         while (iloop == 1) {
             try {
                 // do work   "old way"
-
                 iRet = th.MeasurementTaskDataList();
                 if (iRet != INT_RET_OK) {
-                    if (iRet != NO_TASK_LIST) {
+                    if (iRet == NO_TASK_LIST) {
                         logger.info("Other error iRet = " + iRet);
-                        return OMNIA_DATA_PICK_NOT_OK;
+                        iRet2 = th.DeleteTrashTasksAfterHand();
+                        //iRet2 = th.FillUpTasks();  // just testing RETHINK JIs 08.10 2019 klo 13:42
+                        if (iRet2!=DELETE_TRASH_TASK_OK) {
+                            return OMNIA_DATA_PICK_NOT_OK;
+                        }
+                        logger.info("Successful task delete afterhand continue");
+                        iRet = th.MeasurementTaskDataList();
+                        if (iRet != INT_RET_OK) {
+                            return OMNIA_DATA_PICK_NOT_OK;
+                        }
                     }
                 }
                 ce = th.GetFirstUndoneTaskFromList();
@@ -258,7 +267,7 @@ public class MakeSendJsonOperations {
                     setTaskUnderWork(ce);
                 }
                 if (iRound  >= 2) {
-                    Thread.sleep(5000);
+                    Thread.sleep(getSleep());
                     iRound = 0;
                 }
             } catch (Exception e) {
@@ -273,10 +282,16 @@ public class MakeSendJsonOperations {
     }
    public int   DeleteDoneTaskFromWorkDb() {
        TRPXMeasurementTaskData ce = getTaskUnderWork();
-       int iRet = th.DeleteDoneTaskFromWorkDb(ce);
+       int iRet = th.DeleteDoneTaskFromDb(ce);
+       if (iRet < 0) {   // 0 they have all ready been deleted
+           logger.info("Unsuccessful delete from worktask ce.toString() =  " + ce.toString());
+           logger.info("Unsuccessful delete from worktask iRet = " + iRet);
+           return OMNIA_DATA_PICK_NOT_OK;
+       }
+       iRet = th.DeleteDoneTaskFromWorkDb(ce);
        if (iRet < 1) {
-          logger.info("Unsuccesfull delete from tasklist ce.toString() =  " + ce.toString());
-          logger.info("Unsuccesfull delete from tasklist iRet = " + iRet);
+          logger.info("Unsuccessful delete from Work table ce.toString() =  " + ce.toString());
+          logger.info("Unsuccessful delete from Work table iRet = " + iRet);
           return OMNIA_DATA_PICK_NOT_OK;
       }
       return INT_RET_OK;
